add stage3 as jointspace example, separate codebase to a header and a source file.
