+++ examples/iiqka_moveit_task_constructor_example/kuka_moveit_task_constructor/CMakeLists.txt
@@ -9,9 +9,6 @@ endif()
 find_package(ament_cmake REQUIRED)
 find_package(moveit_task_constructor_core REQUIRED)
 find_package(rclcpp REQUIRED)
-# uncomment the following section in order to fill in
-# further dependencies manually.
-# find_package(<dependency> REQUIRED)
 
 add_executable(kuka_moveit_task_constructor src/kuka_moveit_task_constructor.cpp)
 ament_target_dependencies(kuka_moveit_task_constructor moveit_task_constructor_core rclcpp)
@@ -27,12 +24,7 @@ DESTINATION lib/${PROJECT_NAME})
 
 if(BUILD_TESTING)
 find_package(ament_lint_auto REQUIRED)
-# the following line skips the linter which checks for copyrights
-# uncomment the line when a copyright and license is not present in all source files
-#set(ament_cmake_copyright_FOUND TRUE)
-# the following line skips cpplint (only works in a git repo)
-# uncomment the line when this package is not in a git repo
-#set(ament_cmake_cpplint_FOUND TRUE)
+
 ament_lint_auto_find_test_dependencies()
 endif()
