+++ examples/iiqka_moveit_task_constructor_example/kuka_moveit_task_constructor/include/kuka_moveit_task_constructor/kuka_moveit_task_constructor.hpp
@@ -1,0 +1,40 @@
+#ifndef KUKA_MOVEIT_TASK_CONSTRUCTOR_HPP
+#define KUKA_MOVEIT_TASK_CONSTRUCTOR_HPP
+
+#include <rclcpp/rclcpp.hpp>
+#include <moveit/planning_scene/planning_scene.h>
+#include <moveit/planning_scene_interface/planning_scene_interface.h>
+#include <moveit/task_constructor/task.h>
+#include <moveit/task_constructor/solvers.h>
+#include <moveit/task_constructor/stages.h>
+#if __has_include(<tf2_geometry_msgs/tf2_geometry_msgs.hpp>)
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
+#else
+#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#endif
+#if __has_include(<tf2_eigen/tf2_eigen.hpp>)
+#include <tf2_eigen/tf2_eigen.hpp>
+#else
+#include <tf2_eigen/tf2_eigen.h>
+#endif
+
+namespace mtc = moveit::task_constructor;
+
+class MTCTaskNode
+{
+public:
+  MTCTaskNode(const rclcpp::NodeOptions& options);
+
+  rclcpp::node_interfaces::NodeBaseInterface::SharedPtr getNodeBaseInterface();
+  void doTask();
+  void setupPlanningScene();
+
+private:
+  // Compose an MTC task from a series of stages.
+  mtc::Task createTask();
+  mtc::Task task_;
+  rclcpp::Node::SharedPtr node_;
+  
+};
+
+#endif // KUKA_MOVEIT_TASK_CONSTRUCTOR_HPP
