+++ examples/iiqka_moveit_task_constructor_example/kuka_moveit_task_constructor/src/kuka_moveit_task_constructor.cpp
@@ -1,40 +1,6 @@
-#include <rclcpp/rclcpp.hpp>
-#include <moveit/planning_scene/planning_scene.h>
-#include <moveit/planning_scene_interface/planning_scene_interface.h>
-#include <moveit/task_constructor/task.h>
-#include <moveit/task_constructor/solvers.h>
-#include <moveit/task_constructor/stages.h>
-#if __has_include(<tf2_geometry_msgs/tf2_geometry_msgs.hpp>)
-#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
-#else
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
-#endif
-#if __has_include(<tf2_eigen/tf2_eigen.hpp>)
-#include <tf2_eigen/tf2_eigen.hpp>
-#else
-#include <tf2_eigen/tf2_eigen.h>
-#endif
+#include "kuka_moveit_task_constructor/kuka_moveit_task_constructor.hpp"
 
 static const rclcpp::Logger LOGGER = rclcpp::get_logger("kuka_moveit_task_constructor");
-namespace mtc = moveit::task_constructor;
-
-class MTCTaskNode
-{
-public:
-  MTCTaskNode(const rclcpp::NodeOptions& options);
-
-  rclcpp::node_interfaces::NodeBaseInterface::SharedPtr getNodeBaseInterface();
-
-  void doTask();
-
-  void setupPlanningScene();
-
-private:
-  // Compose an MTC task from a series of stages.
-  mtc::Task createTask();
-  mtc::Task task_;
-  rclcpp::Node::SharedPtr node_;
-};
 
 rclcpp::node_interfaces::NodeBaseInterface::SharedPtr MTCTaskNode::getNodeBaseInterface()
 {
@@ -109,14 +75,8 @@ mtc::Task MTCTaskNode::createTask()
   task.setProperty("group", arm_group_name);
   task.setProperty("ik_frame", eef_frame);
 
-// Disable warnings for this line, as it's a variable that's set but not used in this example
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-but-set-variable"
-  mtc::Stage* current_state_ptr = nullptr;  // Forward current_state on to grasp pose generator
-#pragma GCC diagnostic pop
-
   auto stage_state_current = std::make_unique<mtc::stages::CurrentState>("current");
-  current_state_ptr = stage_state_current.get();
+  mtc::Stage* current_state_ptr = stage_state_current.get();
   task.add(std::move(stage_state_current));
 
   auto sampling_planner = std::make_shared<mtc::solvers::PipelinePlanner>(node_);
@@ -128,19 +88,19 @@ mtc::Task MTCTaskNode::createTask()
   cartesian_planner->setStepSize(.01);
   cartesian_planner->setMinFraction(0.2);
 
-  auto stage1 =
-      std::make_unique<mtc::stages::MoveRelative>("move1", cartesian_planner);
+  // Set stage1 as MoveRelative
+  auto stage1 = std::make_unique<mtc::stages::MoveRelative>("move1", cartesian_planner);
   stage1->properties().set("marker_ns", "move1");
   stage1->properties().set("link", eef_frame);
   stage1->properties().configureInitFrom(mtc::Stage::PARENT, { "group" });
   stage1->setMinMaxDistance(0.0, 0.2);
-  // Set hand forward direction
   geometry_msgs::msg::Vector3Stamped vec;
   vec.header.frame_id = eef_frame;
   vec.vector.z = 1.0;
   stage1->setDirection(vec);
   task.add(std::move(stage1));
 
+  // Set stage2 as MoveTo with cartesian goal
   auto stage2 = std::make_unique<mtc::stages::MoveTo>("move2", cartesian_planner);
   stage2->properties().set("marker_ns", "move2");
   stage2->properties().set("link", eef_frame);
@@ -148,6 +108,7 @@ mtc::Task MTCTaskNode::createTask()
 
   // Set the goal pose for stage2
   geometry_msgs::msg::PoseStamped goalPose;
+  goalPose.header.frame_id = "world";
   goalPose.pose.position.x = 0.4;
   goalPose.pose.position.y = 0.2;
   goalPose.pose.position.z = 0.3;
@@ -157,6 +118,23 @@ mtc::Task MTCTaskNode::createTask()
   goalPose.pose.orientation.w = 1.0;
   stage2->setGoal(goalPose);
   task.add(std::move(stage2));
+
+  // Set stage3 as MoveTo with joint positions
+  auto stage3 = std::make_unique<mtc::stages::MoveTo>("move3", interpolation_planner);
+  stage3->properties().set("marker_ns", "move3");
+  stage3->properties().configureInitFrom(mtc::Stage::PARENT, { "group" });
+
+  // Define the joint positions for the goal
+  std::map<std::string, double> joint_positions = {
+    {"joint_1", 0.0},
+    {"joint_2", -0.5},
+    {"joint_3", 1.0},
+    {"joint_4", -1.5},
+    {"joint_5", 1.0},
+    {"joint_6", 0.5}
+  };
+  stage3->setGoal(joint_positions);
+  task.add(std::move(stage3));
 
   return task;
 }
